requirements.txt
lightgbm==4.1.0
plotly==5.18.0

# Jupyter
jupyter==1.0.0
ipykernel==6.27.1

# Web & API
flask==3.0.0
flask-cors==4.0.0

# FastAPI (API v2)
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.5.0
python-multipart>=0.0.6
apscheduler>=3.10.0
websockets>=12.0

# Utilities
python-dotenv==1.0.0
tqdm==4.66.1

# Production (Fase 7)
gunicorn==21.2.0
schedule==1.2.0
# Core Data Science
pandas>=2.1.0
numpy>=1.26.0        

# Machine Learning
scikit-learn>=1.3.0
xgboost>=2.0.0

# Visualización
matplotlib>=3.7.0
seaborn>=0.12.0

# Utilidades
joblib>=1.3.0
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
tqdm>=4.66.0
